+Added slow_Domain_NH to ice_data_type
  Added an additional domain type element, slow_Domain_NH, to the ice_data_type
for use in redistributing data across processors.  Also eliminated the code
allocating Ice%part_size on the slow ice processors.  Aditionally, modified the
code to update Ice%part_size from the slow ice state to make it conditional on
whether it can be done directly; while this is unneccesary with newer versions
of the coupler code, it is necessary with Verona and earlier couplers.  All
answers are bitwise identical in test cases.