*Avoid NaNs with SIS2_FILLING_FRAZIL=True
  Revised the code that handles the merging of thickness categories when there
is a frazil heat flux to avoid dividing by 0, which could previously have
happened when SIS2_FILLING_FRAZIL was set to true and the part_size for both
categories 0 and 1 was 0.  This can change answers in some circumstances, but it
happens not to have done so for the existing short SIS2 test cases.