Made VERONA_COUPLER a layout param
  Changed the call logging VERONA_COUPLER so that it goes into the
SIS_parameter_doc.layout file, which is allowed to change between
configurations, and not the SIS_parameter_doc.all file, which is not.  Also
removed the default value for VERONA_COUPLER, since this is determined by the
coupler code that is used, so there is effectively no default.  All answers
are bitwise identical, but the SIS_parameter_doc files have changed.