*Reset t_surf to T_freeze ocean when part_size==0
  Previously, t_surf had been set to the freezing point of the ocean whenever
the partition size dropped below 1e-10.  This has now been changed to only reset
t_surf when the partition size is actually 0.  This has been enabled by superior
numerics in SIS2 compared with SIS1, and it avoids having an arbitrary (if
small) value alter the solution in SIS.  This changes answers (slightly) in the
SIS test cases, and new reference solutions have been checked in.