+Added the runtime parameter RECATEGORIZE_ICE
  Added the new runtime parameter RECATEGORIZE_ICE to control whether
adjust_ice_categories is called again a second time toward the end of
ice_transport, to ensure that all of the ice is in the correct thickness
category at the end of the timestep.  By default, all solutions are bitwise
identical, but setting RECATEGORIZE_ICE=True does change answers, and the
SIS_parameter_doc files have changed.