+*Combined CP_WATER and CP_SEAWATER
  For purely historical reasons, there had been separate heat capacities for
water and seawater in various parts of the calculation.  These had been set to
be the same by default, and now they are actually being set identically in the
code using the parameter CP_SEAWATER.  This does also include the removal of one
optional argument for  The answers in the test cases happened not to change, but
new SIS_parameter_doc files were checked in.