(*)Set Ice%t_surf(:,:,1) from SST in Celsius
  Changed how the ocean surface temperature as seen by the atmosphere is set to
use the sea surface temperature in Celsius.  This could change answers at the
level of roundoff (in the test cases it did not do so), but it is a necessary
step for future changes where the ocean provides its temperature to the ice in
Celsius (which is more accurate than converting to and from Kelvin).