Revised the CPU clocks in ice_model.F90
  The FMS CPU clocks must always be used with the same PE-lists, so the clocks
in ice-model have been revised to properly handle the cases where the fast and
slow ice processes use different PEs or the case where they share the same PEs.
When there are separate PEs, there can not be a single clock covering all of the
ice processes, so the next-level PEs become the COMPONENT clocks; otherwise they
are SUBCOMPONENT clocks.  All answers are bitwise identical.