+Modified type redistribute routines
  Modified redistribute_IST_to_IST, redistribute_sOSS_to_sOSS, and
redistribute_FIA_to_FIA to gracefully handle the cases where either the input or
output types are not available because of the processors from which these
routines are being called.  Specifically, the arguments for the types that are
being communicated have been turned into pointers, while several grid arguments
have been made optional.  Calls that worked with the original interface still
work, and all answers are bitwise identical.