*Made exchange_fast_to_slow_ice work with split PEs
  Modified exchange_fast_to_slow_ice and exchange_slow_to_fast_ice to work
gracefully when either the Ice%sCS or Ice%fCS is not associated.  Also moved the
code initializing Ice%rough_mom and related arrays from the slow ice PEs to the
fast ice PEs, where it should have been all along.  When the same PEs are used
for the fast and slow ice, the answers are unchanged, but the new code now can
work if the fast and slow processes run on distinct PEs.