Use symmetric memory for G%gridLonB and G%gridLatB
  G%gridLonB and G%gridLatB use symmetric memory for both the SIS_hor_grid and
the dyn_hor_grid.  This allows these arrays to be used properly as the edge
values for gridLonT and gridLatT in output files.  All answers are bitwise
identical, but the axis labels of output files can change.  This requires the
use of the MOM6 from NOAA-GFDL/MOM6@1b3a51dec8588bbb4771f3856913a04f9e5c50dd or
later.