Set t_skin to 0 over land
  Always set Rad%t_skin to 0 for land points.  Previously it had been set to
either 0 degrees C or the freezing temperature of whatever skin salinity was
being used over land, depending on the PE layout and whether there are any
ice-filled points in a local j-row.  Although this does not change the
solutions, it does change the values of t_skin over land that are in the restart
files.