+Added T_fr_ocn to ocean_sfc_state_type
  Added a new array with the freezing temperature at the ocean's salinity to the
ocean_sfc_state_type and the simple_OSS_type, to reduce the calls to the
freezing point code and simply aspects of the code.  This also is a step toward
making infill_Tskin work for a variety of fields.  All answers are bitwise
identical, but several types and interfaces inside the SIS code have changed.