Added interp_method='none' to vector diagnostics
  FREgrid is unable to properly interpolate vector or tensor fields to different
horizontal grids for general use, so the flag interp_method='none' has been
added to numerous SIS2 vector or tensor field diagnostics, so that they will not
be improperly interpolated.  This is intended to be a temporary fix until the
completion of the CMIP6 runs, after which point FREgrid needs to be fixed to
automatically avoid trying to do things it is incapable of, and not relying on
directives imposed within the model source code. All answers are bitwise
identical, although the metadata for some diagnostics have changed.