+Changed array sizes to match the data domain
  Changed the extents of the temporary arrays in do_update_ice_model_fast and
the array arguments that are passed into sum_top_quantities to match the data
sizes in the SIS_hor_grid_type.  Without this change, there are a set of chksum
calls in do_update_ice_model_fast that were not working properly when the fast
and slow ice states shared the same grid types.  All answers are bitwise
identical.