+Renamed SIS_error_checking to SIS_debugging
  Renamed the module SIS_error_checking to SIS_debugging, which better describes
the intended use of the subroutines that are offered by this modules.  All
answers are bitwise identical, but this requires widespread changes to module
use statements and one subroutine name.  In addition a call to
SIS_debugging_init has been added, which leads to changes in the
SIS_parameter_doc files.  All answers are bitwise identical.