Changed the internal sign convention for drdt
  Changed the sign convention for drdt inside of the SIS2 code to be consistent
with flux_lw, and also renamed drdt to dlwdt.  Several comments have also been
added or changed.  This modificaiton changes the sign of a recently added
diagnostic (and the long name has been modified to reflect this, leading to a
changes in SIS.available_diags files), but all answers are bitwise identical.