(*)Fixed REDO_FAST_ICE_UPDATE=True solutions
  Made several changes the the new code used with REDO_FAST_ICE_UPDATE=True so
that it now conserves heat.  The calculation of the linearized heat fluxes has
also been corrected, which changes diagnostics. This update changes answers if
REDO_FAST_ICE_UPDATE=True, but this is not yet the setting for any of the test
cases.  With this change, the solutions with the various values of
REDO_FAST_ICE_UPDATE have ice area and ice mass that are within a percent of
each other, without any clear systematic difference. This commit lifts the
cautionary caveat against using REDO_FAST_ICE_UPDATE=True that was noted with
SIS2 commit NOAA-GFDL/SIS2@a0d82e4f4cba3af978b8276c2459b9df65403058.