(*)Allocate tracer fluxes in first or second calls
  Added code to allow the number of tracer fluxes in several types to be
determined at either the first or the second call to the copy, redistribute, or
translate routines, and modified several comments to reflect what is going on.
The answers are bitwise identical in the existing test cases, but in new tests
with the embedded ice coupling these changes are essential for the model to run
from a cold start.