+Eliminated FIA%bheat
  The basal heating rate is really a property of the ocean, and not of the fast
ice state, and moreover it is one of the fields that will be updated with the
ocean with synchronous ice coupling, so the bheat element has been eliminated
from the fast ice average type, and is used instead from one of the flavors of
ocean surface state type.  This change required changes to several SIS
interfaces and types, but all answers are bitwise identical.