+Added the new module combined_ice_ocean_driver
  Added a new set of interfaces that are called by couper_main that will step
both the slow sea ice and the ocean model.  This only works for certain
combinations of settings (with error messages in the disallowed configurations),
and it is triggered by setting the variable combined_ice_and_ocean=True in
coupler_nml.  For now the answers are identical to those obtained with
concurrent_ice=True.