(*)Revised compress_ice to avoid underflow issues
  Revised compress_ice to avoid issues when the transfered part size is so small
that there can be underflow.  This attempts to correct a fatal problem that was
arising when we tried to replicate a 1/12-degree HYCOM configuration.  This
modification can change answers in unusual cases (when NaNs would have been
introduced before), but the thresholds for executing the new code have been set
to be so small as to avoid changing answers for any of our existing test cases.