+Added optional gas_flux args to ice_model_init
  Added optional arguments gas_flux and gas_fields_ocn arguments of type
coupler_1d_bc_type to several SIS2 initialization routines, including
ice_model_init.  If these are present, they are used to spawn the types used to
describe additional gas or tracer fluxes and fields at the same time as other
SIS2 arrays are allocated.  There is a new version of the coupler that makes use
of these new arguments, but because they are optional, older versions of the
coupler will work also.  All answers are bitwise identical.