(*)Save gas fluxes to restarts with concurrent ice
  Added calls to save any gas fluxes in the fast-ice-average and
total_surface_flux types to a restart file with concurrent ice coupling, if the
newer version of the coupler, with the optional gas_flux arguments to
ice_model_init, is used.  All answers are bitwise identical, but there is now a
good chance that coupled configurations with gas fluxes and concurrent ice
coupling will reproduce across restarts (this is not well tested by the current
suite of MOM6-examples test cases).  These changes require an updated version of
the coupler_type_reform_rwh branch of the FMS code, which should be made
available more generally as a patch to FMS in or about September 2017.