+Moves XPRT diagnostic into SIS_transport
  Moved the diagnostic XPRT from SIS_dyn_trans into SIS_transport, reducing the
need for cross-module communication.  This diagnostic is identical, but the
order of entries in the SIS_available_diags fiel changes.  Also collected the
code setting up the 2-d version of the ice state used for the dynamics into one
place in the code.  A number of unused variables were eliminated and some OMP
directives were also simplified.  All answers are bitwise identical, but an
optional arguments to post_ice_state_diagnostics was eliminated.