+(*)Consolidated slab_ice advection code
  Consolidated the entirety of the code to do the archaic slab ice advection
inside of slab_ice_advect by the addition of a new optional part_sz argument
that can be set based on whether the tracer (usually ice mass) is 0.  In
addition there is a halo that had been working on category 2 of mH_ice,
reflecting the old SIS1 indexing convention, that was not correct.  This has
been corrected to update mH_ice(:,:,1), following the SIS2 convention.  This
could change answers in cases using SLAB_ICE=True, but as there are no such test
cases (and perhaps this option should be obsoleted altogether) there are no
changes to answers in the MOM6_examples test suite.