+Changed interface to compress_ice
  Elimated three of the arguments to compress_ice and replaced them with an
optional cell_average_state_type argument.  If this is included, the answers
are bitwise identical; if it is excluded they are mathematically identical but
there can be answer changes at the level of roundoff.  Because this argument
is used in the call in the code, all answer are bitwise identical.