+(*)Added the runtime parameter WARSAW_SUM_ORDER
  Added the new runtime parameter WARSAW_SUM_ORDER, which causes answers to
reproduce previous versions if true (the default, unless MERGED_CONTINUITY is
true).  Also introduced the new subroutines set_ocean_top_stress_C2 and
set_ocean_top_stress_B2, which simplify the calculation of the stresses on the
ocean, but with a different order so the answers change at roundoff.  In
addition, if WARSAW_SUM_ORDER is false, the ice coverage is passed to the
dynamics routines in the variable ice_cover, rather than as an implicit
calculation of the complement of IST%part_size(:,:,0).  By default, all answers
are bitwise identical in the MOM6_examples test cases, but there are new entries
in the SIS_parameter_doc files.