+Always allocate IST%u_ice_C and IST%v_ice_C
  Always allocate IST%u_ice_C and IST%v_ice_C, regardless of whether the C-grid
or B-grid spatial discretizations are used.  This allows for significant code
reuse in SIS_dyn_trans and the elimination of two internal variables, so this
may reduce the memory footprint of C-grid configurations.  However, this does
block the determination of whether SIS is using a B-grid or C-grid
discretization by testing whether IST%u_ice_C is allocated, so IST%Cgrid_dyn
must always be used instead.  All answers are bitwise identical.