(*)Corrected a halo update in SIS_dyn_trans
  Corrected the logic controlling a halo update in SIS_dyn_trans, which could
correct a parallel reproduction error when MERGED_CONTINUITY is used with
NSTEPS_ADV > 1.  No existing MOM6-examples test cases test this particular
configuration yet (and this option is relatively new, so it is unlikely to be
used in any cases at all), so all answers in the test cases are bitwise
identical.