+Added ice_cover_transport
  Added a new routine, ice_cover_transport, to the ice_continuity module that
provides an estimate of the ice coverage changes due to advection.  Also added a
new optional argument to SIS_continuity_init and SIS_transport_init (as a
pass-through) to set up the control structure for the ice coverage transport; if
this argument is present the new runtime parameters SIS_COVER_TRANSPORT_SCHEME
and COVER_PPM_VOLUME_BASED_CFL are read and logged, changing the
SIS_parameter_doc files.  All answers are bitwise identical.