+Added INCONSISTENT_COVER_BUG runtime parameter
  Added a new runtime parameter, INCONSISTENT_COVER_BUG, which when false causes
the fractional ice-free coverage to be correctly recalculated at the end of
adjust_ice_categories.  The default for this parameter should be changed to
False and then eventually it should be obsoleted.  A new entry has been added to
the SIS_parameter_doc file.  By default the errors are retained, and all answers
are bitwise identical.