+Obsoleted H_TO_KG_M2
  Obsoleted the run-time parameter H_TO_KG_M2, whose functionality is now
provided by elements of the unit_scale_type.  This changes the SIS_parameter_doc
files, so two minor grammatical errors in these files were changed at the same
time.  All answers are bitwise identical, but the model will not give a fatal
error if H_TO_KG_M2 is specified in the SIS_input files with a value other than
1.0.