+Obsoleted ENTHALPY_UNITS
  Obsoleted the run-time parameter ENTHALPY_UNITS.  The same functionality
is now provided more flexibly and more consistently with MOM6 by setting
Q_RESCALE_POWER.  This change required that a unit_scale_type argument be added
to ice_thermo_init.  All answers are bitwise identical, although it should be
noted that rescaling of the enthalpy units was not working properly before
this change.