+(*)Changed interface to get_sea_surface
  Changed the interface to get_sea_surface to use a hor_index_type to specify
array extents, and to make all of the output variables optional.  These
variables were also renamed to make the external interface clearer.  All answers
in the MOM6-examples test suite are bitwise identical, but they could change at
roundoff due to a different order of arithmetic if a non-zero sea surface
temperature perturbation is used with SPECIFIED_ICE is true.