(*)Refactored ice_model_init
  Refactoring in ice_model_init to facilitate new initialization options and
collect common operations in one place, including one set of halo updates and
one place in the code with the land masking, regardless of the initialization
mode.  Also replaced the use of get_sea_surface with direct calls to
data_override for cases that are not using the specified ice to avoid the
inheritance of some namelist specified limiters with unfortunate defaults to
reproduce ancient sea-ice codes, which could change initial conditions for
some cases.  All answers are bitwise identical in the MOM6-examples test cases.