(*)Corrected ice_state_initialization defaults
  Corrected ice_state_initialization default values for the mode in setting
initial conditions for snow and ice temperature and the default initial
temperature values in uniform_temp and relative_temp modes.  All answers in
the MOM6-examples test suite are bitwise identical, but by default there are
changes in cases that are initialized with non-zero ice amounts but not from
a restart file of some form.  Also, there are changes to the SIS_parameter_doc
files.