+Added TRANSMUTE_ICE option
  Added the option, controlled by the new TRANSMUTE_ICE runtime parameter, to
transform ice and snow directly into seawater with a spatially variable rate,
read from the file specified by TRANSMUTATION_RATE_FILE, as a form of simple
outflow boundary condition on the sea ice.  This commit includes the addition of
two new allocatable arrays in the ice_ocean_flux_type to facilitate monitoring
of pseudo-conservation of heat and salt (bearing in mind that the transmutation
itself causes non-conservation of heat and salt).  There is also a new optional
argument to alloc_ice_ocean_flux.  Also replaced calls to mpp_clock with calls
to cpu_clock from the MOM_cpu_clock module, and replaced calls to file_exist
from FMS_io_mod with calls to file_exists from MOM_io, both in the modules that
are already being modified for this commit.   All answers in the MOM6-examples
test suite are bitwise identical, but there are new entries in the
SIS_parameter_doc files, and there are two new elements in a public type and a
new optional argument in a public subroutine interface.