+Added new runtime parameter ICE_COVER_DISCARD
  Added a new runtime parameter, ICE_COVER_DISCARD, that specifies a fractional
thickness category 1 ice coverage that will be discarded, essentially as a form
of manual underflow.  Other categories roll into category 1 when they have tiny
coverage, so it is only category 1 that requires this.  By default, all answers
are bitwise identical, but there are new entries in the SIS_parameter_doc files.