Use broadcast_domain in share_ice_domains
  Use the new SIS-specific framework level interface broadcast_domain in
share_ice_domains as a substitute for mpp_broadcast_domain, which it wraps. Also
use SIS_framework for the restart interfaces.  These changes should enable any
future infrastructure interface changes without leading to widespread changes
within the SIS2 code or a break in the contuity of solutions.  These are
currently just wrappers of the same subroutines, and all answers are bitwise
identical.