Use init_call argument to ice_data_type_chksum
  Added code to use the optional init_call argument to ice_data_type_chksum,
which was recently omitted when this optional argument was added in an attempt
to partially address SIS2 issue #102.  Without this change, the logical variable
init was being used but not set, so it would have been random whether some of
the chksums were being written.  All solutions are bitwise identical.