+Use SIS-specific restart interfaces
  Replace calls to FMS1 restart routines with calls that go through SIS-specific
interfaces in SIS_framework.F90.  These changes include the addition of 14 new
subroutines or functions in SIS_framework.F90 and the renaming of some of the
FMS1 routines that are available via SIS_framework.  All answers are bitwise
identical, including in restart reproducibility tests for c-grid, b-grid and
coupled test cases.