+Removed various unused arguments
  Removed unused arguments from SIS2 interfaces, especially those related to the
restarts, and changed the type of an argument to SIS_restart_init to the
MOM_domain_type instead of the mpp_domain type to potentially provide
information that can be used for the dimensions variables in the restart files.
All answers are bitwise identical, but there are a number of interface changes.