(*)+Stop ignoring SIS2 namelist entries
  Revised the handling of the SIS2 namelist entries so that they are no longer
ignored, addressing SIS2 issue #96.  This includes modification so that the
behavior when input_filename = 'F' is exactly the same as if it is 'n' if there
are no restart files in the restart_input_dir, or as if it is 'r' if the restart
files are there.  Previously there had been differences in the output files
generated, but the solutions were always if input_filename = 'F'.  The entries
in dirs%restart_output_dir are now used to guide where SIS2 writes its restarts,
whereas this had previously been ignored with the output always going into
"./RESTART/".  As a part of this change, a 5 routines have been imported from
MOM_restart.F90 to help determine when this is a new run in 'F' mode, and there
have been a number of new arguments added to the restart-related interfaces,
both to support these changes, and to align these interfaces with those of
MOM_restart.F90.  Answers can change unless input_filename = 'F', and there can
be changes in which output files are written and the presence of one entry in
the SIS_parameter_doc.all files.