+Revised SIS_restart to avoid direct calls to FMS
  Extensively revised SIS_restart.F90 to read and write restart files using the
MOM_io routines and avoiding the direct use of any FMS routines, which enables
the migration to use the new FMS2 I/O interfaces while still maintaining the
ability to run with the traditional FMS I/O interfaces.  All solutions are
bitwise identical when used with either of the MOM6 config_src/infra directories
(although there are some changes in the output file metadata), and the SIS2 runs
in MOM6-examples with the new restart files reproduce answers when either I/O
via the root PE or parallel I/O are used.  The interfaces that are used from
outside the SIS_restart module are unchanged, although there are some minor
changes to internal interfaces.