+Added 4 new diagnostics to comply with CMOR
  Added 4 new diagnostics to comply with CMOR requirements for units and how the
diagnostics are calculated.  Because the suggested CMOR names were already in
use, they have been added with _CMOR appended to their names.  The four new
diagnostics in question are "siconc_CMOR", "sivol_CMOR", "sisnconc_CMOR" and
"sitemptop_CMOR", and of course there are new entries in the SIS.available_diags
files.  This PR closes SIS2 issues #91, #92, and #93.  In addition by making the
axes in the registration call for "RDG_FRAC" consistent with the dimensionality
of the diagnostic that is written out, this PR might also have addressed SIS2
issue #138.  The new diagnostics have been visually confirmed to be credible,
and all solutions are bitwise identical.