+(*)Fixed the landfast ice algorithm
  Revised the landfast ice discretization so that it is extracting momentum from
the ice and giving it to the land, rather than exchanging it with the ocean.
Also added the new diagnostics FLF_X and FLF_Y to capture the forces due to the
landfast ice, and applied the appropriate dimensional rescaling to the
parameters for the Lemieux landfast ice parameterization.  This will change
answers in cases that use LEMIEUX_LANDFAST = True, and it adds new entries to
the SIS.available_diags files, but all solutions in the MOM6-examples test suite
are bitwise identical.