+Change opt. arg. to set_massless_SIS_tracers
  Changed the interface to set_massless_SIS_tracers by replacing the optional
compute_domain logical argument (which took the opposite sense of what its name
seems to imply) with a new optional integer argument, halos, to specify the
size of the halos that this works on.  Previously calls to this routine from
ice_cat_transport and finish_ice_transport had led to uninitialized data being
used, but this change fixes this problem.  All answers are bitwise identical,
although there is a change to a public interface.