Fix the result of a bad automated merge
  In merging in two separate pull requests that made corrections to similar
lines in the same file, three renamed variables retained their original names,
resulting in SIS2 code that does not compile.  Each of the PRs was fully tested
separately, but for whatever reason their combination was not, and the SIS2 git
repository does not seem to detect certain merge conflicts.  This PR updates
these variable names to their intended values, and the code is once more
compiling and is reproducing the intended outputs.