+Added VEL_UNDERFLOW and STRESS_UNDERFLOW
  Added the new runtime parameters VEL_UNDERFLOW and STRESS_UNDERFLOW to set
minimal values of the velocity and ice-pack lateral stresses that lead to an
underflow of the velocities and stresses to zero.  This new capability can be
needed to give identical results when dimensional consistency tests are carried
out.  Also added the runtime parameter B_DYNAMICS_MIN_MASS to replace a hard
coded value for the B-grid dynamics code.  By default, all answers are bitwise
identical, but there are 2 or 3 new runtime parameters in the SIS_parameter_doc
files.