Icepack area underflow patch  (#184)
New parameter RIDGE_AREA_UNDERFLOW is introduced in order to avoid the occurrence of undetectably small fractional area and negative ice volume following ridging adjustments from Icepack. The default is 0.0, which does not introduce changes in existing runs. A reasonable value would be 10^-26 which would correspond to an Angstrom-scale ice patch within a km-scale grid cell. This addresses #178 and #183