Update SIS2 to use the new IO interface
The legacy file_type and fieldtype handles for field IO are replaced
with the newer MOM IO types (MOM_infra_file, MOM_field), which provide
new FMS-agnostic interfaces to the framework IO.

This also conforms to the new requirement of FMS IO requiring a domain.
(Note that this may be relaxed in the future.)

Thanks to Kate Hedstrom (@kshedstrom) for identifying the issue and
providing the first draft to this fix.