Changes to correct interspersed coupling option
This commit includes the changes to the combined ice ocean driver and a
few other files to (1) allow the combined ice ocean driver to run with a
data atmosphere and (2) correct the interspersed coupling option so that
the information from the updated ocean state is communicated to the
sea-ice.

A new flag has been added to the combined ice ocean driver:
use_intersperse_bug which recovers the answers from the version with the
incorrect coupling. By default, the bug is used. This is because the
Ocean_ice_boundary_type is a nessecary input for the
update_slow_ice_and_ocean subroutine in order to get the correct
coupling.

Changes to ice_model, SIS_dyn_trans, and ice_type are in order to run
the combined ice ocean driver with a data atmosphere. In that case, the
slow and fast sea ice PEs are specified but technically the same PE
lists. In ice_model_init the "concurrent" flag has been renamed to
"split_fast_slow" but it is still called "Concurrent_ice" in the
argument list. A halo update was needed for the sea ice mass when
multiple interspesed cycles are used per coupling time
step(DT_CIOD<dt_cpld).