+Remove rescaling factors from restart files
  Removed the code to account for unit rescaling within the SIS2 restart files.
This rescaling within the restart files has not been used in the code since the
conversion arguments were added to all necessary register_restart_field calls on
about June 2, 2022, as a part of PR #175.  The model will still work with older
restart files provided that they did not use dimensional rescaling, and even if
they did, they can be converted to avoid rescaling with a short run with the
older code that created them.  This commit eliminates the publicly visible
register_unit_conversion_restarts and rescale_fast_to_slow_restart_fields
subroutines, as well as the call to fix_restart_unit_scaling.

  Large sections of rescale_ice_state_restart_fields were eliminated, but this
routine is still needed to handle older (SIS1) restart files, a few of which are
still used for initializing tests of SIS2.

  These changes significantly simplify the code, and they lead to a handful of
constants that are always 1 not being included in the SIS2 restart files. All
answers are bitwise identical, but two publicly visible interfaces have been
eliminated.