Reorder update and Ocn-to-OIB calls
These changes move the ocean to OIB call to the begining of the
interspered loop rather than the end so that it does not need to
be called before the loop starts.
The needed translate OSS to sOSS has also been added to the direct
flux ocn to OIB routine.