Move SIS2 build to $HOME
As with MOM6, this patch temporarily moves the GitLab SIS2 builds from
f5 to $HOME.  In addition to the symbolic link setup used in MOM6, it
also required removal of the SIS2_SRC macro and instead modify the
MOM6-examples/src/SIS2 to point to the PR version.

The CACHE_DIR path also required an update to work on f5.