+Revise SIS2 available diags file to be like MOM6
  Revised the SIS2 diag mediator code that creates the SIS.available_diags file
to follow the same pattern of logging as MOM6.  The specific changes include
moving the module name(s) to their own line, adding static diagnostics to the
variables that are logged in the available_diags file, logging the axes of the
diagnostic, and adding cell metrics to describe how a diagnostic might be
reduced and logging these as well.  All solutions are bitwise identical, but
there are new entries and added information in the SIS.available_diags and
SIS_fast.available_diags files.