+Fix dimensional rescaling of DT_ICE_DYNAMICS
  Added the missing dimensional rescaling factor for DT_ICE_DYNAMICS in
SIS_tracer_advect_init().  This commit could change answers in some case when
dimensional rescaling of time is being tested in SIS2, but because this
particular timestep is only used to determine the maximum number of iterations
for the advection of ice thicknesses and tracers, and the maximum number of
iterations is always at least 1, this appears to have gone undetected heretofore
in the short tests with small CFL numbers that we have used for dimensional
consistency testing in SIS2.  This change includes adding a new unit_scale_type
argument to this same routine.  All answers are bitwise identical when
T_RESCALE_POWER is 0 in SIS_input (as is always the case in production runs),
but there is a new argument to a publicly visible interface.